# apps/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AddBookmarkRequest" : {
        "properties" : {
          "collectionId" : {
            "description" : "생략하면 기본 컬렉션이다 (SPEC-06 §3.5)",
            "format" : "int64",
            "type" : "integer"
          },
          "targetSlug" : {
            "type" : "string"
          },
          "targetType" : {
            "description" : "cocktail · bar · article. Phase 1a 에 실재하는 것은 cocktail 뿐이다",
            "type" : "string"
          }
        },
        "required" : [ "targetSlug", "targetType" ],
        "type" : "object"
      },
...
      "BookmarkItem" : {
        "properties" : {
          "collectionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "nameEn" : {
            "type" : "string"
          },
          "nameKo" : {
            "type" : "string"
          },
          "targetSlug" : {
            "type" : "string"
          },
          "targetType" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "nameEn", "nameKo", "targetSlug", "targetType" ],
        "type" : "object"
      },
...
      "CollectionItem" : {
        "properties" : {
          "bookmarkCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "shareToken" : {
            "type" : "string"
          }
        },
        "required" : [ "bookmarkCount", "id", "name" ],
        "type" : "object"
      },
...
      "CreateCollectionRequest" : {
        "properties" : {
          "name" : {
            "maxLength" : 60,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "SharedCollection" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/SharedItem"
            },
            "type" : "array"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "items", "name" ],
        "type" : "object"
      },
      "SharedItem" : {
        "properties" : {
          "nameEn" : {
            "type" : "string"
          },
          "nameKo" : {
            "type" : "string"
          },
          "targetSlug" : {
            "type" : "string"
          },
          "targetType" : {
            "type" : "string"
          }
        },
        "required" : [ "nameEn", "nameKo", "targetSlug", "targetType" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/api/v1/collections/{shareToken}" : {
      "get" : {
        "description" : "비로그인 조회. 소유자 정보와 내부 id 를 담지 않고, 미발행 항목은 빠진다.",
        "operationId" : "shared",
        "parameters" : [ {
          "in" : "path",
          "name" : "shareToken",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SharedCollection"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "공유된 컬렉션",
        "tags" : [ "shared-collection-controller" ]
      }
    },
...
    "/api/v1/me/bookmarks" : {
      "get" : {
        "description" : "사라지거나 내려간 대상은 빠진다 (다형 참조라 앱이 무결성을 진다).",
        "operationId" : "list",
        "parameters" : [ {
          "description" : "생략하면 전체. 0 이면 기본 컬렉션(collection_id IS NULL)",
          "in" : "query",
          "name" : "collectionId",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/BookmarkItem"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "내 북마크 목록",
        "tags" : [ "bookmark-controller" ]
      },
      "post" : {
        "description" : "targetSlug 로 받는다. 중복은 멱등이다.",
        "operationId" : "add",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddBookmarkRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BookmarkItem"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "북마크 추가",
        "tags" : [ "bookmark-controller" ]
      }
    },
    "/api/v1/me/bookmarks/{id}" : {
      "delete" : {
        "operationId" : "remove",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "summary" : "북마크 삭제",
        "tags" : [ "bookmark-controller" ]
      }
    },
    "/api/v1/me/collections" : {
      "get" : {
        "operationId" : "collections",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CollectionItem"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "내 컬렉션 목록",
        "tags" : [ "bookmark-controller" ]
      },
      "post" : {
        "description" : "생성 시점에 공유 토큰이 함께 발급된다.",
        "operationId" : "createCollection",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateCollectionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionItem"
                }
              }
            },
            "description" : "Created"
          }
        },
        "summary" : "컬렉션 생성",
        "tags" : [ "bookmark-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}